+++ lib_comfyui/async_comfyui_loader.py
@@ -4,7 +4,7 @@ import types
 import sys
 import os
 import runpy
-from lib_comfyui import argv_conversion, custom_extension_injector, webui_resources_sharing, parallel_utils
+from lib_comfyui import argv_conversion, custom_extension_injector, webui_paths, parallel_utils
 
 
 def main(model_attribute_queue, model_apply_queue, shared_opts_queue, comfyui_path):
@@ -17,12 +17,12 @@ def main(model_attribute_queue, model_apply_queue, shared_opts_queue, comfyui_pa
 
 
 def start_comfyui(comfyui_path):
-    folder_paths = webui_resources_sharing.get_folder_paths()
+    folder_paths = webui_paths.get_folder_paths()
 
     sys.path.insert(0, comfyui_path)
     argv_conversion.set_comfyui_argv()
 
-    webui_resources_sharing.share_webui_folder_paths(folder_paths)
+    webui_paths.share_webui_folder_paths(folder_paths)
     custom_extension_injector.register_webui_extensions()
     print('[sd-webui-comfyui]', f'Launching ComfyUI with arguments: {" ".join(sys.argv[1:])}')
     runpy.run_path(os.path.join(comfyui_path, 'main.py'), {}, '__main__')
